# backend/packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: AAAgronegócio Platform API - Farm Listings Marketplace & SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Get published listings with filters and pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 12
        - name: search
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minArea
          in: query
          schema:
            type: number
        - name: maxArea
          in: query
          schema:
            type: number
        - name: propertyType
          in: query
          schema:
            type: string
        - name: aptitude
          in: query
          schema:
            type: string
          description: Texto livre; busca em descrição e campos de atividade da propriedade
        - name: hasWater
          in: query
          schema:
            type: boolean
        - name: hasIrrigation
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, area_asc, area_desc, newest, featured]
            default: featured
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Listings list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsPage"
  /listings/featured:
    get:
      operationId: getFeaturedListings
      tags: [listings]
      summary: Get featured/premium listings for homepage
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 6
      responses:
        "200":
          description: Featured listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListingSummary"
  /listings/map:
    get:
      operationId: getMapListings
      tags: [listings]
      summary: Get listings for map view (lat/lng pins only)
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: propertyType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Map pins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MapPin"
  /listings/{slug}:
    get:
      operationId: getListingBySlug
      tags: [listings]
      summary: Get full listing details by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Listing details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /listings/{slug}/view:
    post:
      operationId: trackListingView
      tags: [listings]
      summary: Track a listing view
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackViewRequest"
      responses:
        "200":
          description: View tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /listings/{slug}/interest:
    post:
      operationId: submitInterest
      tags: [listings]
      summary: Submit buyer interest (generates WhatsApp/email redirect)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterestRequest"
      responses:
        "200":
          description: Interest recorded and contact info returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterestResponse"

  # ─── SELLER LISTINGS ──────────────────────────────────────────────────────────
  /seller/listings:
    get:
      operationId: getSellerListings
      tags: [seller]
      summary: Get seller's own listings
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Seller's listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsPage"
    post:
      operationId: createListing
      tags: [seller]
      summary: Create a new listing (draft)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingRequest"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
  /seller/listings/{id}:
    get:
      operationId: getSellerListing
      tags: [seller]
      summary: Get seller's listing by ID
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
    put:
      operationId: updateListing
      tags: [seller]
      summary: Update a listing
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingRequest"
      responses:
        "200":
          description: Listing updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
    delete:
      operationId: deleteListing
      tags: [seller]
      summary: Delete a listing
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/listings:
    get:
      operationId: adminGetListings
      tags: [admin]
      summary: Get all listings (admin view)
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
        - name: moderationStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminListingsPage"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - error

    # ─── AUTH ───────────────────────────────────────────────────────────────────
...
    ListingSummary:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        price:
          type: number
          nullable: true
        negotiable:
          type: boolean
        totalArea:
          type: number
          nullable: true
        areaUnit:
          type: string
          nullable: true
        pricePerHectare:
          type: number
          nullable: true
        state:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        propertyType:
          type: string
          nullable: true
        isFeatured:
          type: boolean
        isPremium:
          type: boolean
        coverImageUrl:
          type: string
          nullable: true
        thumbnailUrls:
          type: array
          items:
            type: string
        viewCount:
          type: integer
        favoriteCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - title
        - negotiable
        - isFeatured
        - isPremium
        - thumbnailUrls
        - viewCount
        - favoriteCount
        - createdAt
    ListingDetail:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        negotiable:
          type: boolean
        status:
          type: string
          enum: [draft, pending_review, published, rejected, archived]
        moderationStatus:
          type: string
          enum: [pending, approved, rejected]
        moderationNote:
          type: string
          nullable: true
        isFeatured:
          type: boolean
        isPremium:
          type: boolean

        country:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        addressDetails:
          type: string
          nullable: true

        totalArea:
          type: number
          nullable: true
        productiveArea:
          type: number
          nullable: true
        reserveArea:
          type: number
          nullable: true
        areaUnit:
          type: string
          enum: [hectares, alqueires, acres]
          nullable: true
        pricePerHectare:
          type: number
          nullable: true
        pricePerAlqueire:
          type: number
          nullable: true

        propertyType:
          type: string
          nullable: true
        aptidaoAgricola:
          type: boolean
        aptidaoPecuaria:
          type: boolean
        aptidaoMista:
          type: boolean
        hasIrrigation:
          type: boolean
        hasRiver:
          type: boolean
        hasLake:
          type: boolean
        hasAcude:
          type: boolean
        hasWell:
          type: boolean
        hasSpring:
          type: boolean

        infraMainHouse:
          type: boolean
        infraWorkerHouses:
          type: boolean
        infraWarehouse:
          type: boolean
        infraBarn:
          type: boolean
        infraCorral:
          type: boolean
        infraFencing:
          type: boolean
        infraElectricity:
          type: boolean
        infraInternet:
          type: boolean
        infraMachineryShed:
          type: boolean
        infraGrainSilo:
          type: boolean
        infraDairyStructure:
          type: boolean
        infraIrrigationSystem:
          type: boolean
        infraPavedAccess:
          type: boolean
        infraInternalRoads:
          type: boolean
        infraAirstrip:
          type: boolean
        infraOffice:
          type: boolean
        infraSolarPower:
          type: boolean
        infraSecuritySystem:
          type: boolean

        hasCattle:
          type: boolean
        cattleCount:
          type: integer
          nullable: true
        hasHorses:
          type: boolean
        otherAnimals:
          type: string
          nullable: true
        plantedCrops:
          type: string
          nullable: true
        cropDetails:
          type: string
          nullable: true
        productiveActivities:
          type: string
          nullable: true

        bedroomCount:
          type: integer
          nullable: true
        bathroomCount:
          type: integer
          nullable: true
        suiteCount:
          type: integer
          nullable: true
        hasKitchen:
          type: boolean
        hasGarage:
          type: boolean
        hasLeisureArea:
          type: boolean

        deedStatus:
          type: string
          nullable: true
        environmentalLicense:
          type: boolean
        carRegistryInfo:
          type: string
          nullable: true

        media:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"

        seller:
          $ref: "#/components/schemas/SellerPublicInfo"

        viewCount:
          type: integer
        favoriteCount:
          type: integer
        interestCount:
          type: integer

        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - slug
        - title
        - negotiable
        - status
        - moderationStatus
        - isFeatured
        - isPremium
        - aptidaoAgricola
        - aptidaoPecuaria
        - aptidaoMista
        - hasIrrigation
        - hasRiver
        - hasLake
        - hasAcude
        - hasWell
        - hasSpring
        - infraMainHouse
        - infraWorkerHouses
        - infraWarehouse
        - infraBarn
        - infraCorral
        - infraFencing
        - infraElectricity
        - infraInternet
        - infraMachineryShed
        - infraGrainSilo
        - infraDairyStructure
        - infraIrrigationSystem
        - infraPavedAccess
        - infraInternalRoads
        - infraAirstrip
        - infraOffice
        - infraSolarPower
        - infraSecuritySystem
        - hasCattle
        - hasHorses
        - hasKitchen
        - hasGarage
        - hasLeisureArea
        - environmentalLicense
        - media
        - viewCount
        - favoriteCount
        - interestCount
        - createdAt
        - updatedAt
...
    UpdateListingRequest:
      allOf:
        - $ref: "#/components/schemas/CreateListingRequest"
    ListingsPage:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/ListingSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - listings
        - total
        - page
        - limit
        - totalPages
    MapPin:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        price:
          type: number
          nullable: true
        totalArea:
          type: number
          nullable: true
        areaUnit:
          type: string
          nullable: true
        isFeatured:
          type: boolean
        propertyType:
          type: string
          nullable: true
      required:
        - id
        - slug
        - title
        - latitude
        - longitude
        - isFeatured
    TrackViewRequest:
      type: object
      properties:
        source:
          type: string
          nullable: true
        referrer:
          type: string
          nullable: true
        deviceType:
          type: string
          nullable: true
    InterestRequest:
      type: object
      properties:
        interestType:
          type: string
          enum: [info, proposal, visit]
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
      required:
        - interestType
    InterestResponse:
      type: object
      properties:
        whatsappUrl:
          type: string
          nullable: true
        emailUrl:
          type: string
          nullable: true
        message:
          type: string
      required:
        - message
...
    MediaItem:
      type: object
      properties:
        id:
          type: string
        listingId:
          type: string
        url:
          type: string
        mediaType:
          type: string
          enum: [image, video, pdf]
        isPrimary:
          type: boolean
        sortOrder:
          type: integer
        fileName:
          type: string
          nullable: true
        fileSize:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - listingId
        - url
        - mediaType
        - isPrimary
        - sortOrder
        - createdAt
    SellerPublicInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
          nullable: true
        isPremium:
          type: boolean
      required:
        - id
        - name
        - isPremium

    # ─── ANALYTICS ───────────────────────────────────────────────────────────────
...
    AdminListingsPage:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/ListingDetail"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - listings
        - total
        - page
        - limit
        - totalPages
...