# backend/packages/api-zod/src/generated/types/interestResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AAAgronegócio Platform API - Farm Listings Marketplace & SaaS
 * OpenAPI spec version: 1.0.0
 */

export interface InterestResponse {
  whatsappUrl?: string | null;
  emailUrl?: string | null;
  message: string;
}